+++ pkg/test/resources/model/resource.go
@@ -2,14 +2,17 @@ package model
 
 import (
 	core_model "github.com/Kong/kuma/pkg/core/resources/model"
+	"time"
 )
 
 var _ core_model.ResourceMeta = &ResourceMeta{}
 
 type ResourceMeta struct {
-	Mesh    string
-	Name    string
-	Version string
+	Mesh             string
+	Name             string
+	Version          string
+	CreationTime     time.Time
+	ModificationTime time.Time
 }
 
 func (m *ResourceMeta) GetMesh() string {
@@ -20,4 +23,10 @@ func (m *ResourceMeta) GetName() string {
 }
 func (m *ResourceMeta) GetVersion() string {
 	return m.Version
+}
+func (m *ResourceMeta) GetCreationTime() time.Time {
+	return m.CreationTime
+}
+func (m *ResourceMeta) GetModificationTime() time.Time {
+	return m.ModificationTime
 }